+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased](https://github.com/wilfredinni/dinero/compare/0.1.4...master)
 
 - It is no longer posible to campare a `Dinero` instance to objects of other types.
+- Multiplication and division can be performed only with `int`, `float` and `Decimal` types.
 
 ## [0.1.4](https://github.com/wilfredinni/dinero/releases/tag/0.1.4) (2022-11-03)